Decouple artifact download from Presto Dockerfile
Move the artifact (RPM & CLI) download outside of the Presto Dockerfile.
Introduce a `jenkins-release-image.sh` script that can drive the build
and push of new containers. This will replace the inline script
currently used in the Jenkins job.

A pleasant side effect of this change is that the Presto RPM is no
longer left in the resulting container thus reducing the size of the
image from 3.43 GB to 2.41 GB.